Move the writing of the data section to finalize
This was so that we could calculate the record size automatically.

However, this results in the writer being slower as the data value keys
need to be calculated an extra time. Given this, it doesn't make sense
to pursue this minor improvement at this time. Perhaps future
refactoring will make it easier to do this without a performance hit.