Fix control-byte computation at boundary sizes
`writeCtrlByte` writes the left-over value using:

   byte((leftOver >> (8 * i)) & 0xFF)

This means that, after the `leftOver` value has been shifted, it should have a value in the [0,255] range so that the bitmask operation can work properly.

Currently, because the switch case determining `leftOver` is inclusive (<=) rather than inclusive (<), the shifted `leftOver` value can reach the value of 256, making an MMDB file invalid.